requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
pillow==10.0.0
pytesseract==0.3.10
python-multipart==0.0.6
spacy==3.7.0
dateparser==1.1.8
pytz==2024.1
typing_extensions==4.8.0